+++ rest_framework/schemas/views.py
@@ -1,6 +1,5 @@
 from rest_framework import exceptions, renderers
 from rest_framework.response import Response
-from rest_framework.schemas.generators import SchemaGenerator
 from rest_framework.settings import api_settings
 from rest_framework.views import APIView
 
@@ -28,20 +27,3 @@ class SchemaView(APIView):
         if schema is None:
             raise exceptions.PermissionDenied()
         return Response(schema)
-
-
-def get_schema_view(
-        title=None, url=None, description=None, urlconf=None, renderer_classes=None,
-        public=False, patterns=None, generator_class=SchemaGenerator):
-    """
-    Return a schema view.
-    """
-    generator = generator_class(
-        title=title, url=url, description=description,
-        urlconf=urlconf, patterns=patterns,
-    )
-    return SchemaView.as_view(
-        renderer_classes=renderer_classes,
-        schema_generator=generator,
-        public=public,
-    )
